bitkeeper revision 1.1422 (4288fba11spwpuVeAcBTXmV1qAve_Q)
Fix 64-bit build of pygrub. From Jerone Young.

+++ tools/pygrub/src/fsys/ext2/ext2module.c
@@ -55,7 +55,7 @@ static PyObject *
 ext2_file_read (Ext2File *file, PyObject *args)
 {
     int err, size = 0;
-    size_t n, total = 0;
+    unsigned int n, total = 0;
     PyObject * buffer = NULL;
 
     if (file->file == NULL) {
